Adding BASICS documentation.
Trying to draft out a structure of how I would explain this project to
someone with a basic to intermediate understanding of bash, and a passing
understanding of language design.

It will both help me understand these topics better myself, as well as
serve as a good reference for showing this absolute tomfoolery to anyone
else.