Huge restructure. Added ints. Consolidated files.
There was a pretty significant amount of duplicated code across the
data lexer/parser and the query lexer/parser. Was able to consolidate a
pretty significant amount such that it is shared across both.

Was running into some problems as I added an actual int (`declare -i`)
data type. The `get_type()` functions in the interpreter didn't have a
case match for this. That's pretty poor form on my bad. Made it very
frustrating to debug. Need to go back in and clean up all the type case
statements. Should raise type errors if the type is not expected, never
pass silently.