Furthering CLI opts. Changing data flow.
Need to make the flow of data from the data_parser -> query_parser more
explicit, or CLI opts are extremely difficult to do. Going with a
`sed`/`awk` approach, in which you specify some query/script as a
parameter, then pass the data in from stdin or an input file.