Problem when actually deploying. Progdir set wrong.
Fixed by exporting global variable declaring the root of the install
location upon running `shql`. Does make things vaguely less usable if
one wants to run each of the `_parser` scripts as a standalone, but I
think that's becomming less likely to ever be useful, or a realistic
design consideration.